lot of stuff, mainly zuordnung and testing

+++ src/main/java/loc/balsen/kontospring/repositories/KontoRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import loc.balsen.kontospring.data.Konto;
 
-public interface KontoRepository extends JpaRepository<Konto, Long>{
+public interface KontoRepository extends JpaRepository<Konto, Integer>{
 
 	List<Konto> findByKontoGruppeId(Integer id);
 }
